Memory map I/O files and refactor to .c files
