+++ src/lib/trace/include/trace/trace.h
@@ -2,8 +2,6 @@
 
 #include <stddef.h>
 #include <stdint.h>
-#include <stdlib.h>
-#include <string.h>
 
 struct TraceItem {
     uint64_t key;
@@ -30,12 +28,4 @@ struct Trace {
 };
 
 void
-Trace__destroy(struct Trace *const me)
-{
-    if (me == NULL)
-        return;
-
-    // HACK This is an utter hack to get around the const-qualifiers.
-    free((void *)me->trace);
-    *me = (struct Trace){0};
-}
+Trace__destroy(struct Trace *const me);
